pyproject.toml
[tool.poetry]
name = "chess-ai-pip"
version = "0.1.2"
description = ""
authors = ["pacanada <pereirapcanada@gmail.com>"]
readme = "README.md"
packages = [{include = "chess_ai"}]

[tool.poetry.dependencies]
python = "^3.8"
chess-python-pip = "^0.1.8"
pyglet = "^1.5.27"


[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
